# contracts/openapi/scene-service.yaml
...
openapi: 3.0.3
info:
  title: HunterEdge - scene-service
  version: 0.1.0
  description: |
    场景生成服务统一契约（端口 8081，网关前缀 `/api/v1/scene`）。

    **统一响应**：`{code, message, data, request_id, timestamp}`；`code=0` 成功，非 0 为预定义错误码（附录 A）。

    **场景分类**（4.2.1 节）：基础场景（直线巡航 / 弯道行驶 / 跟车行驶 / 变道行驶）、
    交互场景（前车切入 / 行人横穿 / 路口会车 / 施工绕行）、环境场景（雨天 / 夜间 / 雾天 / 强光逆光）、
    极端场景 Corner Case（紧急制动 / 障碍物突现 / 传感器失效 / 多车混行）、自定义场景；
    另有 4.5 节「实车回放」类型。编码值映射见 `x-hunter-scene-types`。

    **场景配置结构**（4.2.2 节，不可更改）：场景元信息（scene_id / scene_name / scene_type /
    description / version / creator / tags）+ map / ego_vehicle / weather / actors[] / events[] /
    success_criteria / duration；字段与落库位置映射见 `x-hunter-scene-config-contract`。

    **状态机**：`draft → published → archived`（`scenes.status`，受控词表）；仅 `draft` 可编辑。

    **鉴权**：请求经网关校验 JWT 后转发，服务侧按 RBAC 资源域 `scene` 校验
    （permissions.resource=scene，action ∈ create/read/update/delete/execute）。

    **限流**：本服务无专属限流（附录 D 仅 3 条接口级限流，不含 scene），继承网关五级限流
    （全局 10000 QPS / 单用户 100 QPS / 单 IP 200 QPS）。

    **性能**：API 响应 P95 ≤ 200ms（13.2 节）；场景列表默认分页 20 条、上限 200 条。
  license:
    name: Proprietary - HunterEdge
servers:
  - url: https://hunter-edge.example.com
    description: 生产环境（经网关 /api/v1/scene 转发）
  - url: http://localhost:8081
    description: 本地开发（直连服务）
...
security:
  - bearerAuth: []
paths:
  /api/v1/scene:
    get:
      tags: [scenes]
      summary: 查询场景列表（分页 + 分类/状态/标签/关键字筛选）
      description: |
        场景库列表。默认按 `create_time DESC` 排序（命中索引
        `idx_scenes_status_type_create_time`）；`tags` 为包含查询（PostgreSQL `@>`，命中 GIN 索引）；
        `keyword` 对 `scene_name`/`description` 做不区分大小写的模糊匹配。
        软删除场景（`deleted_at` 非 NULL）一律不出现在结果中。
      operationId: listScenes
      parameters:
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/PageSizeQuery"
        - name: scene_type
          in: query
          required: false
          description: 场景类型（取值见 `x-hunter-scene-types`）
          schema: { $ref: "#/components/schemas/SceneType" }
        - name: status
          in: query
          required: false
          description: 场景状态（受控词表 draft/published/archived）
          schema: { $ref: "#/components/schemas/SceneStatus" }
        - name: tags
          in: query
          required: false
          description: 标签（可重复传参，多标签为 AND 关系）
          style: form
          explode: true
          schema:
            type: array
            maxItems: 20
            items: { type: string, minLength: 1, maxLength: 32 }
        - name: keyword
          in: query
          required: false
          description: 名称/描述模糊搜索关键字
          schema: { type: string, minLength: 1, maxLength: 64 }
        - name: creator
          in: query
          required: false
          description: 创建者 user_id（数据权限隔离时由服务端强制注入）
          schema: { type: string, format: uuid }
        - name: sort
          in: query
          required: false
          description: 排序字段（白名单，禁止任意字段排序以防注入）
          schema:
            type: string
            enum: [create_time, update_time, scene_name, version]
            default: create_time
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: 场景列表
          headers:
            X-Request-ID: { $ref: "#/components/headers/RequestId" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SceneListResponse" }
        "422": { $ref: "#/components/responses/InvalidParameter" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        "500": { $ref: "#/components/responses/InternalError" }

    post:
      tags: [scenes]
      summary: 创建场景（场景编辑器保存）
      description: |
        创建场景，`status` 固定为 `draft`（不可由请求指定），`creator` 取自网关注入的
        `X-User-Id`，`version` 默认 `1.0.0`。
        `config` 结构必须与 4.2.2 节完全一致（见 `x-hunter-scene-config-contract`）。
        `scene_name` 在同一存活场景内唯一（`uq_scenes_scene_name`），冲突返回 3002。
      operationId: createScene
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SceneCreateRequest" }
      responses:
        "201":
          description: 创建成功（返回完整场景对象，status=draft）
          headers:
            X-Request-ID: { $ref: "#/components/headers/RequestId" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SceneResponse" }
        "422": { $ref: "#/components/responses/InvalidParameter" }
        "409":
          description: 场景名称已存在（code=3002）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /api/v1/scene/{scene_id}:
    get:
      tags: [scenes]
      summary: 查询场景详情
      description: |
        返回场景完整对象（含 4.2.2 结构 `config`）。
        结果按 Redis 契约缓存：`cache:scene:{scene_id}`（String(JSON)，TTL 1 小时），
        任何写操作（PUT / DELETE / publish / duplicate 的源场景无关）后失效。
      operationId: getScene
      parameters:
        - $ref: "#/components/parameters/SceneIdPath"
      responses:
        "200":
          description: 场景详情
          headers:
            X-Request-ID: { $ref: "#/components/headers/RequestId" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SceneResponse" }
        "404": { $ref: "#/components/responses/SceneNotFound" }
        "422": { $ref: "#/components/responses/InvalidParameter" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        "500": { $ref: "#/components/responses/InternalError" }

    put:
      tags: [scenes]
      summary: 更新场景（仅 draft 可编辑）
      description: |
        全量更新场景可编辑字段。规则（4.2 节 + `scenes.status` 受控词表）：
        - 仅 `draft` 状态可编辑；`published` / `archived` 返回 3003
        - `scene_id` / `creator` / `create_time` / `status` / `version` 不可被请求修改
        - `scene_name` 与其它存活场景冲突返回 3002
        - 更新成功后 `update_time` 由触发器/ORM 自动维护
      operationId: updateScene
      parameters:
        - $ref: "#/components/parameters/SceneIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SceneUpdateRequest" }
      responses:
        "200":
          description: 更新后的场景对象
          headers:
            X-Request-ID: { $ref: "#/components/headers/RequestId" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SceneResponse" }
        "404": { $ref: "#/components/responses/SceneNotFound" }
        "409": { $ref: "#/components/responses/SceneStateConflict" }
        "422": { $ref: "#/components/responses/InvalidParameter" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        "500": { $ref: "#/components/responses/InternalError" }

    delete:
      tags: [scenes]
      summary: 删除场景（软删除）
      description: |
        软删除：置 `deleted_at`，不物理删除（保留审计与历史引用）。
        仅 `draft` / `archived` 可删除；`published` 场景必须先归档，直接删除返回 3003
        （⚠ 该规则由契约补充，设计文档只规定「仅 draft 可编辑」，见 `x-hunter-pending-confirmation`）。
      operationId: deleteScene
      parameters:
        - $ref: "#/components/parameters/SceneIdPath"
      responses:
        "200":
          description: 删除结果
          headers:
            X-Request-ID: { $ref: "#/components/headers/RequestId" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SceneDeleteResponse" }
        "404": { $ref: "#/components/responses/SceneNotFound" }
        "409": { $ref: "#/components/responses/SceneStateConflict" }
        "422": { $ref: "#/components/responses/InvalidParameter" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/scene/{scene_id}/duplicate:
    post:
      tags: [scenes]
      summary: 复制场景（生成新草稿）
      description: |
        以源场景 `config` 为模板创建新场景：新 `scene_id`，`status=draft`，
        `creator` = 当前用户，`version` 重置为 `1.0.0`；
        未指定 `new_scene_name` 时按 `<源名称>-copy`（冲突则追加序号）生成，仍冲突返回 3002。
      operationId: duplicateScene
      parameters:
        - $ref: "#/components/parameters/SceneIdPath"
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SceneDuplicateRequest" }
      responses:
        "201":
          description: 复制成功（返回新场景，status=draft）
          headers:
            X-Request-ID: { $ref: "#/components/headers/RequestId" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SceneResponse" }
        "404": { $ref: "#/components/responses/SceneNotFound" }
        "409":
          description: 新场景名称已存在（code=3002）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }
        "422": { $ref: "#/components/responses/InvalidParameter" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/scene/{scene_id}/publish:
    post:
      tags: [scenes]
      summary: 发布场景（draft → published）
      description: |
        发布后场景不可再编辑（仅 `draft` 可编辑），可被 `POST /api/v1/scene/{scene_id}/run`
        下发到 Carla 仿真与导出。
        规则：仅 `draft` 可发布，其它状态返回 3003；发布前服务端再做一次 4.2.2 结构完整性校验。
        ⚠ 版本号递增策略（是否在发布时提升 `version`）设计文档未定义：本契约规定
        `version` 在发布时保持不变，由请求显式指定时才更新（见 `x-hunter-pending-confirmation`）。
      operationId: publishScene
      parameters:
        - $ref: "#/components/parameters/SceneIdPath"
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScenePublishRequest" }
      responses:
        "200":
          description: 发布后的场景对象（status=published）
          headers:
            X-Request-ID: { $ref: "#/components/headers/RequestId" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SceneResponse" }
        "404": { $ref: "#/components/responses/SceneNotFound" }
        "409": { $ref: "#/components/responses/SceneStateConflict" }
        "422": { $ref: "#/components/responses/InvalidParameter" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/scene/export:
    post:
      tags: [export]
      summary: 导出场景（Carla ScenarioRunner XML / OpenSCENARIO 1.2）
      description: |
        按 4.3 节将场景配置导出为仿真可执行文件：

        | format | 产物 | 说明 |
        |--------|------|------|
        | `carla_scenariorunner_xml` | `.xml` | Carla ScenarioRunner 场景文件 |
        | `openscenario_1_2` | `.xosc` | OpenSCENARIO 1.2 标准描述文件（写入 `x-scenario-version`） |

        产物上传 MinIO `hunter-scene-assets`（生命周期：永久），返回预签名下载地址
        （MinIO 契约：下载有效期 15 分钟，支持 Range 分片下载）。
        ⚠ 导出返回形式（文本直返 / MinIO 地址）设计文档未定义，本契约按 MinIO 方案，
        见 `x-hunter-pending-confirmation`。
      operationId: exportScenes
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SceneExportRequest" }
      responses:
        "200":
          description: 导出结果（含预签名下载地址）
          headers:
            X-Request-ID: { $ref: "#/components/headers/RequestId" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SceneExportResponse" }
        "404": { $ref: "#/components/responses/SceneNotFound" }
        "422": { $ref: "#/components/responses/InvalidParameter" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/scene/{scene_id}/run:
    post:
      tags: [simulation]
      summary: 下发场景到 Carla 仿真（4.4 节流程）
      description: |
        按 4.4 节流程下发：① 校验场景配置完整性 → ② 调用 Carla 管理 API 创建仿真实例 →
        ③ 下发场景配置 JSON → ④ Carla 加载地图、生成参与者、设置环境 →
        ⑤ 返回 `sim_instance_id`，场景运行状态置 `running` → ⑥ 实时监控仿真进度 →
        ⑦ 仿真结束回收资源并保存结果（步骤 6/7 的实现方式见 `x-hunter-simulation-flow`）。

        规则：
        - 仅 `published` 场景可下发，其它状态返回 3003
        - Carla 管理 API 不可达/超时返回 5001（依赖服务不可用）
        - 同一场景已有 `running` 实例时按 `sim_config.allow_parallel` 决定复用或 3003
        - 参数覆盖（`param_overrides`）仅覆盖 4.2.2 白名单路径，禁止新增结构字段
      operationId: runScene
      parameters:
        - $ref: "#/components/parameters/SceneIdPath"
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SceneRunRequest" }
      responses:
        "200":
          description: 下发成功（返回仿真实例信息）
          headers:
            X-Request-ID: { $ref: "#/components/headers/RequestId" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SceneRunResponse" }
        "404": { $ref: "#/components/responses/SceneNotFound" }
        "409": { $ref: "#/components/responses/SceneStateConflict" }
        "422": { $ref: "#/components/responses/InvalidParameter" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Access Token（2 小时，网关签发）。载荷：`sub`(user_id) / `username` / `roles` / `permissions` /
        `iat` / `exp` / `iss` / `jti`。网关校验后注入 `X-User-Id`、`X-Roles`、`X-Trace-Id`，
        服务侧据此执行 RBAC（`permissions.resource=scene`，action ∈ create/read/update/delete/execute）
        与数据权限（按 creator/项目隔离）。
    deviceCertificate:
      type: mutualTLS
      description: |
        车辆设备认证：X.509 双向 TLS，证书 CommonName = vehicle_id。
        本服务无车端 REST 端点，此处仅声明全系统安全约定（车-云数据通道为 Kafka SASL_SSL）。
  parameters:
    SceneIdPath:
      name: scene_id
      in: path
      required: true
      description: 场景 ID（UUID，对应 scene_svc.scenes.scene_id）
      schema: { type: string, format: uuid }
    PageQuery:
      name: page
      in: query
      required: false
      description: 页码（≥1）
      schema: { type: integer, minimum: 1, default: 1 }
    PageSizeQuery:
      name: page_size
      in: query
      required: false
      description: 每页条数（1-200，与 hunter_common BaseRepository.paginate 上限一致）
      schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
...
  headers:
    RequestId:
      description: 链路追踪 ID（与请求头 `X-Request-ID` 一致）
      schema: { type: string, maxLength: 64 }
  responses:
    Unauthenticated:
      description: 未认证（code=1001，Token 缺失或无效）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiResponse" }
          example:
            code: 1001
            message: 未认证
            data: null
            request_id: 8f9c2b7a-1d3e-4f5a-9b6c-0d1e2f3a4b5c
            timestamp: 1724035200
...
    Forbidden:
      description: 无权限（code=1002，RBAC 资源域 scene 权限不足）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiResponse" }
          example:
            code: 1002
            message: 无权限
            data: null
            request_id: 8f9c2b7a-1d3e-4f5a-9b6c-0d1e2f3a4b5c
            timestamp: 1724035200
    InvalidParameter:
      description: 参数错误（code=2001，请求体/查询参数校验失败；HTTP 422，见 x-hunter-error-status-map）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiResponse" }
          example:
            code: 2001
            message: 参数错误
            data: null
            request_id: 8f9c2b7a-1d3e-4f5a-9b6c-0d1e2f3a4b5c
            timestamp: 1724035200
...
    SceneNotFound:
      description: 场景不存在（code=3001，含已软删除场景）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiResponse" }
          example:
            code: 3001
            message: 资源不存在
            data: null
            request_id: 8f9c2b7a-1d3e-4f5a-9b6c-0d1e2f3a4b5c
            timestamp: 1724035200
...
    SceneStateConflict:
      description: |
        资源状态冲突（code=3003）：非 draft 编辑 / published 删除 / 非 published 下发 /
        已有运行中仿真实例且未允许并行。
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiResponse" }
          example:
            code: 3003
            message: 资源状态冲突
            data: null
            request_id: 8f9c2b7a-1d3e-4f5a-9b6c-0d1e2f3a4b5c
            timestamp: 1724035200
    ServiceUnavailable:
      description: 服务不可用（code=5001）：PostgreSQL / Redis / MinIO / Carla 管理 API 不可达
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiResponse" }
          example:
            code: 5001
            message: 服务不可用
            data: null
            request_id: 8f9c2b7a-1d3e-4f5a-9b6c-0d1e2f3a4b5c
            timestamp: 1724035200
    InternalError:
      description: 服务器内部错误（code=5000，未预期异常）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiResponse" }
          example:
            code: 5000
            message: 服务器内部错误
            data: null
            request_id: 8f9c2b7a-1d3e-4f5a-9b6c-0d1e2f3a4b5c
            timestamp: 1724035200
  schemas:
...
    ApiResponse:
      type: object
      description: 统一响应格式（所有业务响应/错误响应均为该结构）
      required: [code, message, data, request_id, timestamp]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
          example: success
        data:
          nullable: true
          description: 业务数据；错误时为 null 或错误上下文
        request_id:
          type: string
          description: 请求追踪 ID（等于响应头 X-Request-ID）
          example: 8f9c2b7a-1d3e-4f5a-9b6c-0d1e2f3a4b5c
        timestamp:
          type: integer
          format: int64
          description: Unix epoch 秒
          example: 1724035200
...
    SceneStatus:
      type: string
      description: |
        场景状态受控词表（contracts/database/enums.md 第 6 节，DB CHECK 约束同源）。
        `draft` 可编辑；`published` 可下发/导出；`archived` 只读可删。
      enum: [draft, published, archived]
    SceneType:
      type: string
      description: |
        场景类型（4.2.1 分类体系的叶子场景；"分类 → 类型"映射见 `x-hunter-scene-types`）。
        ⚠ 编码值由本契约提出（设计文档未给出编码），需人工确认后落库，禁止前端硬编码分支。
      enum:
        - straight_cruise       # 基础场景：直线巡航
        - curve_driving         # 基础场景：弯道行驶
        - car_following         # 基础场景：跟车行驶
        - lane_change           # 基础场景：变道行驶
        - cut_in                # 交互场景：前车切入
        - pedestrian_crossing   # 交互场景：行人横穿
        - intersection_meeting  # 交互场景：路口会车
        - construction_detour   # 交互场景：施工绕行
        - rainy                 # 环境场景：雨天
        - night                 # 环境场景：夜间
        - foggy                 # 环境场景：雾天
        - backlight             # 环境场景：强光逆光
        - emergency_braking     # 极端场景：紧急制动
        - obstacle_appearance   # 极端场景：障碍物突现
        - sensor_failure        # 极端场景：传感器失效
        - multi_vehicle_mixed   # 极端场景：多车混行
        - custom                # 自定义场景
        - real_vehicle_replay   # 实车回放（4.5 节实车场景自动提取产物）
...
    SceneConfig:
      type: object
      description: |
        场景仿真配置（4.2.2 节结构中的仿真相关段，落库 scenes.config_json）。
        7 个字段全部必填，与 x-hunter-scene-config-contract 声明一致。
      required: [map, ego_vehicle, weather, actors, events, success_criteria, duration]
      properties:
        map: { $ref: "#/components/schemas/SceneMap" }
        ego_vehicle: { $ref: "#/components/schemas/EgoVehicle" }
        weather: { $ref: "#/components/schemas/Weather" }
        actors:
          type: array
          maxItems: 100
          description: 参与者列表（可为空数组）
          items: { $ref: "#/components/schemas/Actor" }
        events:
          type: array
          maxItems: 100
          description: 事件列表（可为空数组）
          items: { $ref: "#/components/schemas/SceneEvent" }
        success_criteria: { $ref: "#/components/schemas/SuccessCriteria" }
        duration: { type: number, exclusiveMinimum: 0, description: "场景时长（s）；⚠ 上限未定义，下发前校验" }
...
    Scene:
      description: 场景完整对象 = SceneMeta（元信息，4.2.2 元信息段）+ config（仿真配置段）
      allOf:
        - $ref: "#/components/schemas/SceneMeta"
        - type: object
          required: [config]
          properties:
            config: { $ref: "#/components/schemas/SceneConfig" }
...
    SceneListData:
      type: object
      required: [items, total, page, page_size]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Scene" }
        total: { type: integer, minimum: 0, description: "满足筛选条件的总数（软删除数据不计入）" }
        page: { type: integer, minimum: 1 }
        page_size: { type: integer, minimum: 1, maximum: 200 }
...
    SceneDeleteData:
      type: object
      required: [scene_id, deleted]
      properties:
        scene_id: { type: string, format: uuid }
        deleted: { type: boolean, description: "软删除结果（true=deleted_at 已写入）" }
...
    SceneResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          required: [data]
          properties:
            data: { $ref: "#/components/schemas/Scene" }
    SceneListResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          required: [data]
          properties:
            data: { $ref: "#/components/schemas/SceneListData" }
...
    SceneDeleteResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          required: [data]
          properties:
            data: { $ref: "#/components/schemas/SceneDeleteData" }
    SceneExportResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          required: [data]
          properties:
            data: { $ref: "#/components/schemas/SceneExportData" }
    SceneRunResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          required: [data]
          properties:
            data: { $ref: "#/components/schemas/SceneRunData" }

    # ---------- 请求体 ----------
    SceneCreateRequest:
      type: object
      description: |
        创建场景请求：元信息（写入 scenes 列）+ config（写入 config_json，4.2.2 结构）。
        status / creator / version / create_time 由服务端生成，不接收客户端传值。
      required: [scene_name, scene_type, config]
      additionalProperties: false
      properties:
        scene_name: { type: string, minLength: 1, maxLength: 128 }
        scene_type: { $ref: "#/components/schemas/SceneType" }
        description: { type: string, nullable: true, maxLength: 1024 }
        tags:
          type: array
          maxItems: 20
          default: []
          items: { type: string, minLength: 1, maxLength: 32 }
        config: { $ref: "#/components/schemas/SceneConfig" }
    SceneUpdateRequest:
      type: object
      description: |
        更新场景请求（PUT 全量语义）。元信息与 config 必须整体提交，
        运行中/已发布场景返回 3003；status / version / creator 不可改。
      required: [scene_name, scene_type, config]
      additionalProperties: false
      properties:
        scene_name: { type: string, minLength: 1, maxLength: 128 }
        scene_type: { $ref: "#/components/schemas/SceneType" }
        description: { type: string, nullable: true, maxLength: 1024 }
        tags:
          type: array
          maxItems: 20
          default: []
          items: { type: string, minLength: 1, maxLength: 32 }
        config: { $ref: "#/components/schemas/SceneConfig" }
    SceneDuplicateRequest:
      type: object
      description: 复制场景请求（可省略请求体）
      additionalProperties: false
      properties:
        new_scene_name:
          type: string
          nullable: true
          minLength: 1
          maxLength: 128
          description: "新场景名称；省略时按 <源名称>-copy 生成（冲突追加序号）"
    ScenePublishRequest:
      type: object
      description: |
        发布场景请求（可省略请求体）。
        ⚠ 版本号策略：本契约规定仅在显式传 version 时更新 scenes.version，否则保持不变。
      additionalProperties: false
      properties:
        version:
          type: string
          nullable: true
          pattern: "^\\d+\\.\\d+\\.\\d+$"
          description: "发布版本（可选，语义化版本；需 ≥ 当前版本）"
...
    SceneExportRequest:
      type: object
      description: |
        导出请求：支持批量导出（一次一个文件，多场景合并写入同一导出文件集）。
        ⚠ 批量语义（单文件内多场景 vs 多文件打包）设计文档未定义，见 x-hunter-pending-confirmation。
      required: [scene_ids, format]
      additionalProperties: false
      properties:
        scene_ids:
          type: array
          minItems: 1
          maxItems: 50
          items: { type: string, format: uuid }
        format: { $ref: "#/components/schemas/SceneExportFormat" }
    SceneRunRequest:
      type: object
      description: |
        场景下发请求（可省略请求体）。
        param_overrides 仅允许覆盖 4.2.2 结构中的白名单路径（如 ego_vehicle.initial_speed、
        weather.*、duration），服务端以白名单校验后深合并；禁止覆盖 map.map_id。
      additionalProperties: false
      properties:
        param_overrides:
          type: object
          additionalProperties: true
          description: "参数化覆盖（白名单路径深合并；键为点分路径）"
        sim_config:
          type: object
          description: "仿真运行配置（⚠ 设计文档未定义配置项，见 x-hunter-pending-confirmation）"
          additionalProperties: false
          properties:
            allow_parallel:
              type: boolean
              default: false
              description: "允许同一场景并行运行多个实例（默认 false，避免资源争抢）"


# =====================================================================
# 扩展字段：机器可校验契约（由 services/scene-service/app/tests/test_openapi_contract.py 强制校验）
# =====================================================================
...